refactor: use `src` directory
