+++ source-code/website/src/pages/editor/@host/@owner/@repository/Layout.tsx
@@ -13,7 +13,7 @@ import { subSeconds, isAfter } from "date-fns";
 import { currentPageContext } from "@src/renderer/state.js";
 import { showToast } from "@src/components/Toast.jsx";
 import { Layout as RootLayout } from "@src/pages/Layout.jsx";
-import { useLocalStorage } from "@src/services/local-storage/LocalStorageProvider.jsx";
+import { useLocalStorage } from "@src/services/local-storage/index.js";
 import type { EditorRouteParams } from "./types.js";
 import { onFork } from "@src/services/github/index.js";
 import { navigate } from "vite-plugin-ssr/client/router";
