+++ source-code/website/src/services/auth/index.ts
@@ -1,2 +1,4 @@
 export { SignInDialog } from "./src/components/SignInDialog.jsx";
 export { onSignOut } from "./src/onSignOut.js";
+// a telefunc automatically runs on the server, thus can be used on the client
+export { getUserInfo } from "./src/implementation.telefunc.js";
