+++ source-code/website/src/services/auth/src/implementation.telefunc.ts
@@ -1,4 +1,4 @@
-import type { LocalStorageSchema } from "@src/services/local-storage/schema.js";
+import type { LocalStorageSchema } from "@src/services/local-storage/index.js";
 import { getContext } from "telefunc";
 
 /**
